Refatorado tabelas e DAOS em pacotes, refatorado configurção de banco de dados e inicialização de tabelas
