+++ src/main/kotlin/newDocument/persistence/tables/person/OfficialTable.kt
@@ -1,9 +1,9 @@
-package newDocument.persistence.tables
+package newDocument.persistence.tables.person
 
+import newDocument.persistence.tables.NotaryTable
 import newDocument.person.Sex
 import org.jetbrains.exposed.dao.id.EntityID
 import org.jetbrains.exposed.dao.id.IdTable
-import org.jetbrains.exposed.dao.id.UUIDTable
 import org.jetbrains.exposed.sql.Column
 import java.util.*
