+++ src/main/kotlin/newDocument/service/DocumentService.kt
@@ -15,28 +15,16 @@ import newDocument.handlers.person.PersonHandler
 import newDocument.persistence.DatabaseInitializer
 import org.jetbrains.exposed.sql.Database
 import org.jetbrains.exposed.sql.transactions.TransactionManager
+import kotlin.system.exitProcess
 
 fun main() {
     try {
-        val host = System.getenv("DATABASE_HOST")
-        val port = System.getenv("DATABASE_PORT")
-        val database = System.getenv("DATABASE_NAME")
-        val user = System.getenv("DATABASE_USER")
-        val password = System.getenv("DATABASE_PASSWORD")
-        val url = "jdbc:mysql://$host:$port/$database?verifyServerCertificate=false&useSSL=true"
-        val db = Database.connect(
-            url = url,
-            driver = "com.mysql.jdbc.Driver",
-            user = user,
-            password = password,
-        )
-
-        TransactionManager.defaultDatabase = db
+        DatabaseInitializer.loadConfigurations()
+        DatabaseInitializer.initialize()
     } catch (e:Exception) {
-        println(e.message)
+        e.printStackTrace()
+        exitProcess(1)
     }
-
-    DatabaseInitializer.initialize()
 
     val notaryHandler = NotaryHandler()
     val personHandler = PersonHandler()
